requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.1
django-cors-headers==3.11.0
django-toolbelt==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
urllib3==1.26.9
